changed validator structure
